Removed DropdownToggle mixin
