+++ addon/components/base/bs-dropdown/button.js
@@ -1,5 +1,5 @@
 import Button from 'ember-bootstrap/components/bs-button';
-import DropdownToggle from 'ember-bootstrap/mixins/dropdown-toggle';
+import { schedule } from '@ember/runloop';
 
 /**
  Button component with that can act as a dropdown toggler.
@@ -9,7 +9,20 @@ import DropdownToggle from 'ember-bootstrap/mixins/dropdown-toggle';
  @class DropdownButton
  @namespace Components
  @extends Components.Button
- @uses Mixins.DropdownToggle
  @public
  */
-export default Button.extend(DropdownToggle);
+export default Button.extend({
+  classNames: ['dropdown-toggle'],
+
+  didReceiveAttrs() {
+    this._super(...arguments);
+    let dropdown = this.get('dropdown');
+    if (dropdown) {
+      schedule('actions', this, function() {
+        if (!this.get('isDestroyed')) {
+          dropdown.set('toggle', this);
+        }
+      });
+    }
+  }
+});
